Move artist and podcaster items to separate files
