+++ lib/features/artists/artists_page.dart
@@ -3,9 +3,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:rating/app/core/enums.dart';
 import 'package:rating/app/injection_container.dart';
 import 'package:rating/features/artists/cubit/artists_cubit.dart';
-import 'package:rating/features/details/artists/pages/artists_details_page.dart';
+import 'package:rating/features/artists/widgets/artist_item.dart';
 import 'package:rating/features/widgets/search_box2.dart';
-import 'package:rating/model/artist_model.dart';
 
 class ArtistsPage extends StatelessWidget {
   const ArtistsPage({super.key});
@@ -101,7 +100,7 @@ class ArtistsPage extends StatelessWidget {
                     },
                     itemBuilder: (context, index) {
                       final artistModel = state.artistModel[index];
-                      return _ArtistItemWidget(
+                      return ArtistItemWidget(
                         artistModel: artistModel,
                       );
                     },
@@ -135,83 +134,6 @@ class ArtistsPage extends StatelessWidget {
             );
           },
         ),
-      ),
-    );
-  }
-}
-
-class _ArtistItemWidget extends StatelessWidget {
-  const _ArtistItemWidget({
-    Key? key,
-    required this.artistModel,
-  }) : super(key: key);
-
-  final ArtistModel artistModel;
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => ArtistsDetailsPage(
-              artistModel: artistModel,
-            ),
-          ),
-        );
-      },
-      child: Container(
-        width: 250,
-        height: 300,
-        decoration: ShapeDecoration(
-          color: Colors.white,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(30),
-          ),
-          shadows: const [
-            BoxShadow(color: Colors.white, blurRadius: 10, spreadRadius: 0),
-          ],
-        ),
-        child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              const SizedBox(
-                height: 10,
-              ),
-              CircleAvatar(
-                radius: 90,
-                backgroundImage: AssetImage(
-                    'assets/graphics/artists_avatars/artist${artistModel.place}.png'),
-              ),
-              const SizedBox(
-                height: 7,
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 18.0),
-                child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      Text(
-                        '#${artistModel.place.toString()}',
-                        style: const TextStyle(
-                          fontSize: 30,
-                          fontFamily: 'Jura',
-                          fontWeight: FontWeight.w700,
-                        ),
-                      ),
-                      Text(
-                        artistModel.name,
-                        style: const TextStyle(
-                          fontSize: 20,
-                          fontFamily: 'Jura',
-                          fontWeight: FontWeight.w500,
-                        ),
-                      ),
-                    ]),
-              ),
-            ]),
       ),
     );
   }
